working

+++ src/main/kotlin/com/lgaertner/euler/usecase/LargestPrimeFactorUseCase.kt
@@ -3,7 +3,7 @@ package com.lgaertner.euler.usecase
 import com.lgaertner.euler.math.PrimeFactors
 
 class LargestPrimeFactorUseCase {
-    fun execute(number: Int): Int {
+    fun execute(number: Long): Long {
         val largestPrimeFactor = PrimeFactors(number).values().max()
         return largestPrimeFactor
     }
